+++ gdbt/provider/__init__.py
@@ -13,3 +13,4 @@
 import gdbt.provider.grafana  # noqa
 import gdbt.provider.prometheus  # noqa
 import gdbt.provider.consul  # noqa
+import gdbt.provider.file  # noqa