+++ gdbt/state/plan.py
@@ -18,25 +18,25 @@ class Plan:
         resources = self.diff.current.resources
         outcomes = self.diff.outcomes(providers)
         for key in sorted(outcomes.keys(), reverse=True):
-            with halo.Halo(text="Loading", spinner="dots") as spinner:
+            with halo.Halo(text="Applying changes", spinner="dots") as spinner:
                 if outcomes[key].action == "create":
+                    spinner.text = f"Creating resource {key}"
                     resource = self.diff.desired.serialize(providers)[key]
                     resource_cls = type(deserialize.deserialize(gdbt.resource.resource.Resource, resource))
                     resource.pop("kind", None)
                     resource_created = resource_cls.create(providers=providers, **resource)
-                    spinner.text = f"Creating resource {key}"
                     resources.update({key: resource_created})
                     spinner.succeed(f"Created resource {key}")
                 elif outcomes[key].action == "delete":
-                    resource = self.diff.current.resources[key]
                     spinner.text = f"Deleting resource {key}"
+                    resource = self.diff.current.resources[key]
                     resource.delete(providers)
                     resources.pop(key, None)
                     spinner.succeed(f"Deleted resource {key}")
                 else:
+                    spinner.text = f"Updating resource {key}"
                     resource = self.diff.current.resources[key]
                     resource_new = self.diff.desired.resources[key]
-                    spinner.text = f"Updating resource {key}"
                     resource.update(resource_new.model, providers)
                     resources.update({key: resource_new})
                     spinner.succeed(f"Updated resource {key}")