+++ CHANGELOG.md
@@ -4,11 +4,23 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
-## [1.0.0] - 2020-11-03
+## [1.0.1] - 2020-12-07
+
+### Added
+
+- File provider for state
+
+### Fixed
+
+- Minor UI/UX fixes and tweaks
+- Fixed `id` field not being ignored
+
+## [1.0.0] - 2020-12-07
 
 ### Added
 
 - Initial version
 
-[unreleased]: https://github.com/SupersonicAds/spotcli/compare/v1.0.0...HEAD
+[unreleased]: https://github.com/SupersonicAds/spotcli/compare/v1.0.1...HEAD
+[1.0.1]: https://github.com/SupersonicAds/spotcli/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/SupersonicAds/spotcli/compare/2c07324...v1.0.0